gameState.tsのswitch文をgameStateUtilsに移動
